requirements.txt
# Google APIs
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Scheduling
schedule>=1.2.0
python-dateutil>=2.8.2

# Data validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0
